requirements.txt
argcomplete==3.3.0
arrow==1.3.0
asgiref==3.8.1
click==8.1.7
crispy-tailwind==1.0.3
Django==5.0.6
django-crispy-forms==2.2
djangorestframework==3.15.2
Faker==22.2.0
mysql-connector-python==9.0.0
numpy==1.26.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pipx==1.5.0
platformdirs==4.2.2
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
userpath==1.9.2
WMI==1.4.9
